+++ coloring/colorizer/export_test.go
@@ -1,0 +1,9 @@
+// Copyright 2021 Nelson Germán Ghezzi. All rights reserved.
+// Use of this source code is governed by a MIT license that
+// can be found in the LICENSE file.
+
+package colorizer
+
+func Private(o Option) {
+	o.private()
+}